% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_fleet}
\alias{appstream_create_fleet}
\title{Creates a fleet}
\usage{
appstream_create_fleet(Name, ImageName, ImageArn, InstanceType,
  FleetType, ComputeCapacity, VpcConfig, MaxUserDurationInSeconds,
  DisconnectTimeoutInSeconds, Description, DisplayName,
  EnableDefaultInternetAccess, DomainJoinInfo, Tags)
}
\arguments{
\item{Name}{[required] A unique name for the fleet.}

\item{ImageName}{The name of the image used to create the fleet.}

\item{ImageArn}{The ARN of the public, private, or shared image to use.}

\item{InstanceType}{[required] The instance type to use when launching fleet instances. The following
instance types are available:
\itemize{
\item stream.standard.medium
\item stream.standard.large
\item stream.compute.large
\item stream.compute.xlarge
\item stream.compute.2xlarge
\item stream.compute.4xlarge
\item stream.compute.8xlarge
\item stream.memory.large
\item stream.memory.xlarge
\item stream.memory.2xlarge
\item stream.memory.4xlarge
\item stream.memory.8xlarge
\item stream.graphics-design.large
\item stream.graphics-design.xlarge
\item stream.graphics-design.2xlarge
\item stream.graphics-design.4xlarge
\item stream.graphics-desktop.2xlarge
\item stream.graphics-pro.4xlarge
\item stream.graphics-pro.8xlarge
\item stream.graphics-pro.16xlarge
}}

\item{FleetType}{The fleet type.

ALWAYS\_ON

:   Provides users with instant-on access to their apps. You are charged
for all running instances in your fleet, even if no users are
streaming apps.

ON\_DEMAND

:   Provide users with access to applications after they connect, which
takes one to two minutes. You are charged for instance streaming
when users are connected and a small hourly fee for instances that
are not streaming apps.}

\item{ComputeCapacity}{[required] The desired capacity for the fleet.}

\item{VpcConfig}{The VPC configuration for the fleet.}

\item{MaxUserDurationInSeconds}{The maximum time that a streaming session can run, in seconds. Specify a
value between 600 and 360000.}

\item{DisconnectTimeoutInSeconds}{The time after disconnection when a session is considered to have ended,
in seconds. If a user who was disconnected reconnects within this time
interval, the user is connected to their previous session. Specify a
value between 60 and 360000.}

\item{Description}{The description to display.}

\item{DisplayName}{The fleet name to display.}

\item{EnableDefaultInternetAccess}{Enables or disables default internet access for the fleet.}

\item{DomainJoinInfo}{The name of the directory and organizational unit (OU) to use to join
the fleet to a Microsoft Active Directory domain.}

\item{Tags}{The tags to associate with the fleet. A tag is a key-value pair, and the
value is optional. For example, Environment=Test. If you do not specify
a value, Environment=.

If you do not specify a value, the value is set to an empty string.

For more information, see \href{https://docs.aws.amazon.com/appstream2/latest/developerguide/tagging-basic.html}{Tagging Your Resources}
in the \emph{Amazon AppStream 2.0 Developer Guide}.}
}
\description{
Creates a fleet. A fleet consists of streaming instances that run a
specified image.
}
\section{Request syntax}{
\preformatted{appstream$create_fleet(
  Name = "string",
  ImageName = "string",
  ImageArn = "string",
  InstanceType = "string",
  FleetType = "ALWAYS_ON"|"ON_DEMAND",
  ComputeCapacity = list(
    DesiredInstances = 123
  ),
  VpcConfig = list(
    SubnetIds = list(
      "string"
    ),
    SecurityGroupIds = list(
      "string"
    )
  ),
  MaxUserDurationInSeconds = 123,
  DisconnectTimeoutInSeconds = 123,
  Description = "string",
  DisplayName = "string",
  EnableDefaultInternetAccess = TRUE|FALSE,
  DomainJoinInfo = list(
    DirectoryName = "string",
    OrganizationalUnitDistinguishedName = "string"
  ),
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
