% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_image}
\alias{appstream_delete_image}
\title{Deletes the specified image}
\usage{
appstream_delete_image(Name)
}
\arguments{
\item{Name}{[required] The name of the image.}
}
\description{
Deletes the specified image. You cannot delete an image when it is in
use. After you delete an image, you cannot provision new capacity using
the image.
}
\section{Request syntax}{
\preformatted{appstream$delete_image(
  Name = "string"
)
}
}

\keyword{internal}
