% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_delete_user}
\alias{appstream_delete_user}
\title{Deletes a user from the user pool}
\usage{
appstream_delete_user(UserName, AuthenticationType)
}
\arguments{
\item{UserName}{[required] The email address of the user.}

\item{AuthenticationType}{[required] The authentication type for the user. You must specify USERPOOL.}
}
\description{
Deletes a user from the user pool.
}
\section{Request syntax}{
\preformatted{appstream$delete_user(
  UserName = "string",
  AuthenticationType = "API"|"SAML"|"USERPOOL"
)
}
}

\keyword{internal}
