% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_fleet}
\alias{appstream_start_fleet}
\title{Starts the specified fleet}
\usage{
appstream_start_fleet(Name)
}
\arguments{
\item{Name}{[required] The name of the fleet.}
}
\description{
Starts the specified fleet.
}
\section{Request syntax}{
\preformatted{appstream$start_fleet(
  Name = "string"
)
}
}

\keyword{internal}
